requirements.txt
colorlog>=6.8.2
homeassistant>=2024.6.0
pip>=21.3.1
ruff>=0.6.1
pyExoyOne>=0.1.2
